requirements.txt
google-generativeai
fastapi
uvicorn
openai
python-dotenv
pdfplumber
faiss-cpu
numpy

pydantic
